Make the script less error prone
First of all, the script now checks if Chaotic-AUR keyring and mirrorlist are downloaded correctly, and terminates execution otherwise.

Secondly, pacman skips installing all packages even if only few of them can not be installed (for example, missing from the repos), which is not an ideal behaviour for us. So before installing the packages the script now checks which of them are present in the repos and skips only those which are missing.